chore: rename withdrawal modal file to follow rule
