+++ src/common/components/shared/modals/withdrawal-error.tsx
