+++ src/common/components/shared/modals/withdrawal-success.tsx
